Valamis 3.0

+++ valamis-scorm-lesson/src/main/scala/com/arcusys/valamis/lesson/scorm/model/tracking/AttemptData.scala
@@ -4,5 +4,5 @@ class AttemptData(
   //val id: Int,
   val dataKey: String,
   val dataValue: Option[String],
-  val attemptID: Int,
+  val attemptID: Long,
   val activityID: String)
